requirements.txt
fastapi
uvicorn[standard]
python-dotenv
slack-bolt
slack-sdk
reportlab
matplotlib